requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy>=2.0.0
psycopg2>=2.9.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv>=1.0.0
google-generativeai==0.3.2
python-multipart==0.0.9
bcrypt==4.1.2
# Add any other dependencies your project uses 